# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Management API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the Task Management application."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/content/flags/stats": {
      "get": {
        "summary": "Get statistics on flagged content",
        "responses": {
          "200": {
            "description": "Flagged content statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalFlaggedPosts": {
                      "type": "integer",
                      "description": "Total number of flagged posts"
                    },
                    "totalFlaggedUsers": {
                      "type": "integer",
                      "description": "Total number of flagged users"
                    },
                    "mostCommonFlagReason": {
                      "type": "string",
                      "description": "The most common reason for flagging content"
                    },
                    "flaggedContentByCategory": {
                      "type": "object",
                      "description": "A breakdown of flagged content by category",
                      "additionalProperties": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
...
}