# src/api/v1/controllers/moderationController.ts
import { Request, Response } from "express";



/**
 * @swagger
 * /api/v1/moderation/post/{id}:
 *   get:
 *     summary: "Retrieve a post by ID"
 *     description: "Fetch a specific post by its unique ID."
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         description: The ID of the post
 *         schema:
 *           type: string
 *           example: "1234"
 *     responses:
 *       200:
 *         description: "Successfully retrieved post."
 *         content:
 *           application/json:
 *             example:
 *               id: "1234"
 *               content: "This is a post"
 *               userId: "5678"
 *       400:
 *         description: "Invalid ID supplied"
 *         content:
 *           application/json:
 *             example:
 *               error: "Invalid ID"
 *               details: "The ID must be a valid post ID."
 *       404:
 *         description: "Post not found"
 *         content:
 *           application/json:
 *             example:
 *               error: "Post not found"
 *       500:
 *         description: "Internal server error"
 *         content:
 *           application/json:
 *             example:
 *               error: "Server error"
 */
export const getPostById = (req: Request, res: Response): void => {
	res.status(200).json({
		message: "Post retrieved successfully",
		data: {
			id: req.params.id,
			content: "Sample post content here...",
			author: "Author ID or Name",
			isFlagged: false,
			createdAt: "2023-10-01T12:34:56Z",
			updatedAt: "2023-10-02T08:00:00Z",
		},
	});
};
/**
 * @swagger
 * /api/v1/moderation/user/{id}/profile:
 *   get:
 *     summary: "Retrieve a user's profile"
 *     description: "Fetch a user’s profile by their unique ID."
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         description: The ID of the user
 *         schema:
 *           type: string
 *           example: "5678"
 *     responses:
 *       200:
 *         description: "Successfully retrieved user profile."
 *         content:
 *           application/json:
 *             example:
 *               id: "5678"
 *               username: "john_doe"
 *               flaggedPosts: 5
 *       400:
 *         description: "Invalid ID supplied"
 *         content:
 *           application/json:
 *             example:
 *               error: "Invalid user ID"
 *               details: "The user ID must be valid."
 *       404:
 *         description: "User not found"
 *         content:
 *           application/json:
 *             example:
 *               error: "User not found"
 */
export const getUserProfile = (req: Request, res: Response): void => {
	res.status(200).json({
		message: "User profile retrieved successfully",
		data: {
			id: req.params.id,
			username: "sampleUser123",
			bio: "This is a sample bio for the user profile.",
			isFlagged: false,
			joinedAt: "2023-01-15T09:00:00Z",
			postsCount: 45,
		},
	});
};
/**
 * @swagger
 * /content/flags/stats:
 *   get:
 *     summary: Get statistics on flagged content
 *     responses:
 *       200:
 *         description: Flagged content statistics
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 totalFlaggedPosts:
 *                   type: integer
 *                   description: Total number of flagged posts
 *                 totalFlaggedUsers:
 *                   type: integer
 *                   description: Total number of flagged users
 *                 mostCommonFlagReason:
 *                   type: string
 *                   description: The most common reason for flagging content
 *                 flaggedContentByCategory:
 *                   type: object
 *                   description: A breakdown of flagged content by category
 *                   additionalProperties:
 *                     type: integer
 */

/**
 * Retrieve statistics on flagged content
 * @param req - Express request object
 * @param res - Express response object
 */
export const getFlaggedContentStats = (req: Request, res: Response): void => {
	res.status(200).json({
		message: "Flagged content statistics",
		data: {
			totalFlaggedPosts: 120,
			totalFlaggedUsers: 15,
			mostCommonFlagReason: "Spam",
			flaggedContentByCategory: {
				spam: 75,
				hateSpeech: 30,
				inappropriateContent: 15,
			},
		},
	});
};
/**
 * @swagger
 * /api/v1/moderation/post/{id}/moderate:
 *   post:
 *     summary: "Moderate a post"
 *     description: "Perform moderation actions on a post (e.g., flag or remove)."
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         description: The ID of the post to moderate
 *         schema:
 *           type: string
 *           example: "1234"
 *       - in: body
 *         name: action
 *         required: true
 *         description: Action to take on the post (e.g., flag, remove).
 *         schema:
 *           type: object
 *           properties:
 *             action:
 *               type: string
 *               enum: [flag, remove]
 *               example: "flag"
 *     responses:
 *       200:
 *         description: "Post moderated successfully."
 *         content:
 *           application/json:
 *             example:
 *               message: "Post flagged successfully"
 *       400:
 *         description: "Invalid action supplied"
 *         content:
 *           application/json:
 *             example:
 *               error: "Invalid action"
 *               details: "The action must be either 'flag' or 'remove'."
 *       404:
 *         description: "Post not found"
 *         content:
 *           application/json:
 *             example:
 *               error: "Post not found"
 *       500:
 *         description: "Internal server error"
 *         content:
 *           application/json:
 *             example:
 *               error: "Server error"
 */
export const moderatePost = (req: Request, res: Response): void => {
	res.status(200).json({
		message: "Post moderated successfully",
		data: {
			id: req.params.id,
			status: "Moderated",
			actionTaken: "Content flagged and hidden",
			moderatedAt: new Date().toISOString(),
		},
	});
};

/**
 * @swagger
 * /api/v1/moderation/user/{id}/flag:
 *   post:
 *     summary: "Flag a user"
 *     description: "Flag a user for inappropriate behavior."
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         description: The ID of the user to flag
 *         schema:
 *           type: string
 *           example: "5678"
 *     responses:
 *       200:
 *         description: "User flagged successfully."
 *         content:
 *           application/json:
 *             example:
 *               message: "User flagged successfully"
 *       400:
 *         description: "Invalid user ID"
 *         content:
 *           application/json:
 *             example:
 *               error: "Invalid user ID"
 *       500:
 *         description: "Internal server error"
 *         content:
 *           application/json:
 *             example:
 *               error: "Server error"
 */
export const flagUser = (req: Request, res: Response): void => {
	res.status(200).json({
		message: "User flagged successfully",
		data: {
			userId: req.params.id,
			reason: req.body.reason || "Spam",
			flaggedAt: new Date().toISOString(),
		},
	});
};